Improved Job Matching
Split the processing of job listings input from the user via source
CSVs into two steps:   at the beginning, we load all CSV jobs that are
marked as "Inactive" or not interested by the user and use that as the
start job list for the run.  At the end, we update the run's full job
list with the inputted active jobs, letting the user's active job
records take precedence and write their notes to those rows.