Fixes for plugins that hit slow jobs sites
Switched to using CURL instead of the file_get_contents for better
control & reliability on slow sites
Changed markup for dupes to just be <dupe>XXX</dupe>.
Changed unknown companies to be "company-unknown"
Improved CL plugin to get the majority of results returned for a
search.  Still misses some if there are multiple with the same title
and date.