Refactored include file path handling and output file/folders
The path the user specifies still is used to output the single results
file, but now all other CSVs (regular subviews of the jobs data and the
debug CSVs) are all put in a single subfolder with the same base name
as the jobs file itself.  ".\Myjobs.CSV" would be a file at
".\Myjobs.CSV" and a folder of files at ".\Myjobs\.*"

Refactored how require_once is done and removed reference to external
library from project.  This fixes alternate instances of functions
getting inadvertently used instead of the right ones.

Moved Linkify,php to lib for consistency.
Moved runJobs.php to /main.

Added check on ClassJobsSitePlugin to verify that it was configured
correctly before any function can be called.  This prevents use of the
function if the developer has not specified the flags and output
folders.
Removed ClassJobsSitePluginNoActualSite subclass since only the
JobRunner class was using it.  It uses ClassJobsSitePlugin but
overrides the constructor and destructors.

Added new ADVANCED_TEXT_CLEANUP option to strScrub that keeps
capitalization in place, but removes extra space, punctuation, etc.
Used for Company name column in results.