Major update of config; now uses ScooperLib
- Now requires ScooperLib, the common library of classes & functions
for Scooper and Jobs_Scooper.  ScooperLib is loaded via Composer.

- Added new Location and Keywords optional values to searches in the
INI.  The base_url_format is automatically set at the plugin class
level so that the user does not have to configure it.  You can still
set a url_format option value in the INI for any plugin's search to use
that value instead of the standard one in the plugin.  The new Location
& Keywords options are supported by all plugins except Amazon,
CareerBuilder, eBay, Disney, EmploymentGuide, Expedia, Google, and
SimplyHired.

- Results CSV file now includes only updated job listings and any
active ones in the list.  It no longer includes inactive listings that
were not updated today.

- Updated ReadMe and example_config.ini for new INI search options

- Added initial basic support for Ubuntu Linux.  Note that any
ApplesCripts will fail on any platform other than Mac OS X.  This means
that some plugins (Amazon, Google, Geekwire, DotJobs) will fail.  Also,
email notifications will not be sent.

- Removed old, unnecessary _getURLfromBase_ instance from
PluginGlassdoor.

- Removed old, unnecessary __dumpGlobalArray__ function

- Minor update to output results table

- Initial attempt to include any errors from the run in the final
results text displayed in the log and emailed to the user.