Fixed:  searches with just url_format values (no keywords/location) fail to return results
-- Search configs that could support keywords but didn't have a
***KEYWORDS*** tag in the base URL would not get added to the plugin as
searches to be run.  We now check if the ***KEYWORDS*** tag exists in
the base URL for that search first.  If there is no tag, we add the
search to be run.  If there is a tag, then we do the normal keywords
checking and configuration.   Also made the same change for
***LOCATION*** tags in the URL.

-- Fixed:  Mashable, Seattle Times, CareerCast plugins fail to load
results

-- Also added support for a CareerCast plugin to specify extra
parameters such as "view = national network" as a prefix on the base
URL. Mashable now uses that to get all national results, not just
Mashable-only ads.

-- Finishes fix for Issue #14 - Plugins that have search settings with
url_format settings fail to return jobs