Keyword sets are applied to all searches, even if they don't support keywords in the URL
-- Fixes issue #29 Plugins that return all results from a site should
still be run if keyword settings are set.  Previously, if the plugin
did not support keywords in the URL, we wouldn't add it as a search for
a keyword set.  Now, we do add it as a search when there is a keyword
set.  The results from that plugin get returned and then filtered on
title after download.

--  Moved class closing log entry to be debug only

-- Improved logging of email addresses loaded from config.ini

-- Improved email notification logging: now includes to / from
addresses