ClassConfig now handles the logic for expanding keyword & location sets into separate searches
Moved logic from the runner class to the Config class so that it's done
up-front and contained in the same class as the rest of the
keyword/location loading logic.

All the input file determination logic and the filter exclusion data
loading is also done in the ConfigClass.  Basically, now we know
everything except the jobs themselves before we start a run.  We can
start marking things now as soon as a plugin gets its data rather than
waiting to the end because everything is pre-loaded.

Still need to remove passing the sets around at the runner and multi
class levels.