Major refactor of how keyword and location sets become searches
-- This is the second part of a refactor that was started within
introduction of the Config class.  Processing of keyword sets and
location sets into the correct searches now happens at the beginning as
part of configuration.  Each search is given a copy of it's necessary
location and keyword set to do what it might need, but creation of
searches for those sets now happens up front.

-- Changed array keys to be standard names for "key" and "name"