Perf:  now caching job results by URL; no longer caching objects
- Perf:  now caching job results by URL; no longer caching objects
The previous objects cache model turned out to run out of memory when
it tried to process a lot of searches.  Instead, we now cache the
resulting full, unfiltered job listing set by URL and re-use that when
possible.

- Fixed:  all jobs returned for keyword search if site always returns
all jobs.  Now that search is skipped if we cannot do keyword matching
through that site plugin.

- Misc log cleanup

- Removed duplicate calls to setKeywords since FinalizeSearch takes
care of it