Perf:  Added Rudimentary File-based Results Caching
This is the first step in getting to file-based caching of interim
results someday in order massively reduce the memory overhead we
currently have.  This change is the first small step along that path
and is mainly a performance optimization for the case where the app is
run more than once per day.

* Job search results are cached with the following cache key
components:  pull date, search name and initial search URL for page 1 /
item 1.
* All job listings for that search and any subsequent pages are cached
to disk and used on future runs that hit the same cache key (basically
any runs in the same day.)

Known limitations:
* Cache does not clean up after itself.
* Cache is not yet useful in any scenario other than running more than
per day.  (Super handy to have for development & debugging however!)