Commented out "Wait for Class" support for dynamic sites
If we didn’t find the class, the plugin would return an error for the
search.  Usually however, it just meant that there were no actual job
matches, not that an error occurred.  Because it returned an error, the
no results found weren’t ever cached.

For now, the minor benefits of this minor wait for class piece aren’t
worth the headaches caused by the multiple cache misses.