* Massively simplified the use of additionalFlags on plugins by splitting out the pagination settings into a separate member variable (paginationType) and using easier-to-understand string values for the types. * Removed concept of C__JOB_SETTINGS_GET_ALL_JOBS_UNFILTERED.  It ended up being exactly the same as C__JOB_KEYWORD_URL_PARAMETER_NOT_SUPPORTED.  Therefore just replaced C__JOB_SETTINGS_GET_ALL_JOBS_UNFILTERED with C__JOB_KEYWORD_URL_PARAMETER_NOT_SUPPORTED to further simplify the number of flags.
Now that most of the flags are auto-set based on the presence or absence of other settings (see commit #b6f735a) and pagination to it's own field, nearly every plugin was able to be simplified and had most, if not all, of its flag settings removed.   This makes the plugin easier to write and to understand as well as reduces the likely hood of incorrectly set flags.   All plugins were reflected to take advantage of this new setup.