Added new "[user.<name>]" config section for setting all the facts about a user's alert run
Moved search keywords and search locations to be per user settings.

Replaced UserSearches & UserKeywordSets in the data model with a new UserSearchPair object that combines a location and a keyword.  Keywords are now single atoms and we don't bother trying to work with sets of sets of keywords anymore.

Added new autogenerated ID column to user_search_site_run table and set unique index on the former primary key columns