Rewrote Geolocation caching to use PhpFastCache instead of a custom array cache pool.   Global object is now under JSCOOP_CACHES.  Currently using the Files driver and storing the cache under <output> in the 'caches' directory
Reduced data stored in the cache to just the lookup values and the GeoLocationId value, not copies of the whole GeoLocation structure.

Removed cache warm up execution for the time being since the cache will reload itself as it needs locations.

Moved the LocationManager cache to be under JSCOOP_CACHES

Geocoding a location now directly returns a hydrated GeoLocation object if it was able to successfully geocode.