Fixed (further):  is_job_match and Is_excluded not always set correctly for job matches
Finally gave up on trying to use Propel's Array column type here.  It was causing more confusion / issues than it was worth.  Switched the three keyword array columns to be instead varchar columns in the database which store a tokenized list of strings (token = "_||_").  This makes the checks for "hasMatch" much simpler and cleaner:  is it empty or does it have characters?   On set, the Propel data objects handles transformation between strings or arrays to the database values.  On get, we always return the list of strings as an array, leaving the logic for joining & splitting the DB strings in a single small codepath which is easy to understand and maintain.

Updated related code to use the new Propel field names for each migrated DB column.

UserJobMatch:  also set default "false" values for boolean columns on this table rather than having strangely null "boolean" values.